refactor(baseline): split baseline.rs into a submodule directory (#68)
src/baseline.rs was 1274 LOC, over the 1000-LOC soft cap in the module
size budget. Split it into src/baseline/ mirroring the run/ and
enrich/maintainer/ splits from the #57 series: a thin mod.rs that owns
the Baseline/BaselineEntry types and parsing core, plus apply.rs
(suppression pass), mutate.rs (baseline-file writing), comment.rs
(directive parsing), and tests.rs.

Pure mechanical move, no behavior change. The public surface stays
reachable at crate::baseline::* and pub mod baseline; in lib.rs is
unchanged. doc_kind widened from private to pub(super) since it is now
shared across mutate.rs and the tests.

Closes #68.
