+++ src/baseline/mutate.rs
@@ -1,0 +1,140 @@
+//! Baseline file mutation: append advisory IDs to a baseline's
+//! `suppressed_advisories` array (the `bomdrift baseline add` flow).
+
+use std::path::Path;
+
+use anyhow::{Context, Result};
+
+use crate::clock;
+
+use super::doc_kind;
+
+/// Append an advisory ID to a baseline file's `suppressed_advisories` array.
+/// Used by the v0.5 `bomdrift baseline add <id>` subcommand and by the
+/// `comment-suppress` sub-action.
+///
+/// Behavior
+///
+/// - **File doesn't exist**: a new baseline is created at `path` with
+///   `{"schema_version": 1, "suppressed_advisories": ["<id>"]}`. The
+///   parent directory is created if missing.
+/// - **File exists but isn't valid JSON**: returns an error. We don't
+///   want to silently overwrite hand-edited baselines on a typo.
+/// - **File exists and is valid JSON**: parses, appends `id` to the
+///   `suppressed_advisories` array (idempotent — skips if already
+///   present), and writes back atomically via temp-file + rename.
+/// - **`id` is empty or whitespace**: returns an error rather than
+///   adding a noise entry.
+///
+/// The write preserves any existing fields in the JSON document
+/// (`changes`, `enrichment`, comments via `_note`, etc.) so this can
+/// safely be run against a baseline generated by `bomdrift diff
+/// --output json`.
+pub fn add_suppression(path: &Path, id: &str) -> Result<AddOutcome> {
+    add_suppression_full(path, id, None, None)
+}
+
+/// Like [`add_suppression`] but accepts optional `expires` (YYYY-MM-DD) and
+/// `reason` fields. When either is provided, the new entry is written in
+/// the v0.8 object form `{id, expires?, reason?}`; existing string-form
+/// entries elsewhere in the array are left untouched.
+pub fn add_suppression_full(
+    path: &Path,
+    id: &str,
+    expires: Option<&str>,
+    reason: Option<&str>,
+) -> Result<AddOutcome> {
+    let trimmed = id.trim();
+    if trimmed.is_empty() {
+        anyhow::bail!("advisory id must not be empty");
+    }
+    if let Some(s) = expires {
+        clock::parse_ymd(s).with_context(|| format!("invalid --expires {s:?}"))?;
+    }
+
+    let mut doc: serde_json::Value = if path.exists() {
+        let body = std::fs::read_to_string(path)
+            .with_context(|| format!("reading baseline file: {}", path.display()))?;
+        if body.trim().is_empty() {
+            serde_json::json!({})
+        } else {
+            serde_json::from_str(&body)
+                .with_context(|| format!("parsing baseline JSON: {}", path.display()))?
+        }
+    } else {
+        serde_json::json!({})
+    };
+
+    if !doc.is_object() {
+        anyhow::bail!(
+            "baseline file root must be a JSON object, found: {}",
+            doc_kind(&doc)
+        );
+    }
+
+    #[allow(
+        clippy::expect_used,
+        reason = "invariant: is_object() check above guarantees Value::Object so as_object_mut() returns Some"
+    )]
+    let obj = doc
+        .as_object_mut()
+        .expect("invariant: is_object() check above guarantees Value::Object");
+    obj.entry("schema_version")
+        .or_insert(serde_json::Value::from(1u64));
+
+    let arr = obj
+        .entry("suppressed_advisories")
+        .or_insert_with(|| serde_json::Value::Array(Vec::new()));
+    let arr = arr
+        .as_array_mut()
+        .ok_or_else(|| anyhow::anyhow!("baseline `suppressed_advisories` field is not an array"))?;
+
+    let already_present = arr.iter().any(|v| match v {
+        serde_json::Value::String(s) => s == trimmed,
+        serde_json::Value::Object(o) => o.get("id").and_then(|x| x.as_str()) == Some(trimmed),
+        _ => false,
+    });
+    if already_present {
+        return Ok(AddOutcome::AlreadyPresent);
+    }
+    if expires.is_some() || reason.is_some() {
+        let mut entry = serde_json::Map::new();
+        entry.insert("id".into(), serde_json::Value::String(trimmed.to_string()));
+        if let Some(s) = expires {
+            entry.insert("expires".into(), serde_json::Value::String(s.to_string()));
+        }
+        if let Some(s) = reason {
+            entry.insert("reason".into(), serde_json::Value::String(s.to_string()));
+        }
+        arr.push(serde_json::Value::Object(entry));
+    } else {
+        arr.push(serde_json::Value::String(trimmed.to_string()));
+    }
+
+    if let Some(parent) = path.parent()
+        && !parent.as_os_str().is_empty()
+    {
+        std::fs::create_dir_all(parent)
+            .with_context(|| format!("creating parent dir: {}", parent.display()))?;
+    }
+
+    // Atomic temp-file + rename, mirroring src/refresh.rs's pattern.
+    let tmp_path = path.with_extension("json.tmp");
+    let serialized = serde_json::to_string_pretty(&doc).context("serializing baseline JSON")?;
+    std::fs::write(&tmp_path, serialized)
+        .with_context(|| format!("writing temp baseline: {}", tmp_path.display()))?;
+    std::fs::rename(&tmp_path, path)
+        .with_context(|| format!("renaming temp baseline to: {}", path.display()))?;
+
+    Ok(AddOutcome::Added)
+}
+
+/// Result of [`add_suppression`]. `Added` means the file was modified;
+/// `AlreadyPresent` means the ID was already in the suppressed list and
+/// no write happened. Both are non-error paths — callers can use this to
+/// emit a precise log message.
+#[derive(Debug, PartialEq, Eq)]
+pub enum AddOutcome {
+    Added,
+    AlreadyPresent,
+}
